# [logger] Optimize log formatting performance (35-72% faster)