# What does this implement/fix?

Optimizes logger header formatting for 13x speedup by replacing `snprintf()` with manual string construction. This improves overall logging performance by 35-72% depending on message complexity.

## Performance Improvements

**Header-only performance (isolated):**
- Simple messages: 82.1 ns → 6.3 ns (**13.0x speedup, 92.3% faster**)
- With thread names: 76.8 ns → 10.8 ns (**7.1x speedup, 85.9% faster**)

**Full logging performance (realistic, including vsnprintf body + footer):**
- Sensor logs with floats: 234.6 ns → 153.4 ns (**1.53x speedup, 34.6% faster**)
- Simple messages: 119.3 ns → 34.0 ns (**3.51x speedup, 71.5% faster**)
- Complex multi-arg logs: 181.2 ns → 103.5 ns (**1.75x speedup, 42.9% faster**)

**Memory impact:**
- ESP32: +228 bytes flash (+0.012%)
- ESP32-C3: +340 bytes flash (+0.019%)
- ESP8266: +112 bytes flash (+0.011%), **-128 bytes RAM (-0.16%)**

The RAM savings on ESP8266 come from storing only ANSI color digit characters instead of full escape sequences, eliminating string array overhead.

## Technical Details

### What Changed

1. Replaced the `snprintf()` call in `write_header_to_buffer_()` with manual buffer construction using direct memory operations and character-by-character assembly
2. Optimized ANSI color code storage by storing only the color digit character (e.g., `'1'` for red) instead of full escape sequences (e.g., `"\033[0;31m"`)
3. Dynamic ANSI sequence construction using `write_ansi_color_for_level()` which builds the escape sequence on-the-fly

### Why This Is Faster

- **No format string parsing** - `snprintf()` must parse `%s`, `%03u`, etc. at runtime
- **Direct memory operations** - Uses simple `memcpy()` and direct character writes
- **Better compiler optimization** - Manual construction allows compiler to optimize with `-Os`
- **Compile-time constants** - Uses `constexpr` for ANSI color digit chars and level letter chars
- **Reduced memory footprint** - Char arrays instead of string arrays save RAM (especially on ESP8266)

## Why This Matters

- Logging is one of the most frequently executed code paths in ESPHome
- When investigating issues, adding logging can trigger loop blocking warnings and alter timing, making bugs harder to track down. Faster logging reduces this "observer effect"

## Real-World Validation

Tested on ESP32-C3 with actual device config dump logging:
- **Console (direct):** 201ms → 143ms (**58ms faster, 29% improvement**)
- **Over network:** 253ms → 240ms (13ms faster, 5% improvement)

The console results closely match the benchmark predictions. The network results show the optimization still provides measurable benefits even when I/O overhead dominates.

## Benchmarks

Comprehensive benchmarks are included in `logger_optimization_benchmarks.zip`:
- `benchmark_logger_noinline.cpp` - Header-only performance measurement
- `benchmark_full_logging.cpp` - Full logging pipeline with realistic sensor messages

Both benchmarks use `-Os` (optimize for size) as ESPHome does in production.

**Note:** Benchmarks were run on a host machine. Actual ESP hardware (especially ESP8266 at 80MHz) will be significantly slower in absolute terms, but the relative speedup ratio (13x for header, 1.5-3.5x for full logging) remains valid.

## Types of changes

- [x] Code quality improvements to existing code or addition of tests
- [x] Performance optimization (non-breaking)

**Related issue or feature (if applicable):**

- N/A

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- N/A (internal optimization, no user-visible changes)

## Test Environment

- [ ] ESP32
- [x] ESP32 IDF
- [x] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx
- [ ] nRF52840

Tested on ESP32-C3 (IDF) and ESP8266. Benchmarks validated on host. The optimization is platform-agnostic and benefits all targets.

## Example entry for `config.yaml`:

```yaml
# No configuration changes required - this is an internal optimization
# All existing logger configurations work exactly as before

logger:
  level: DEBUG
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [x] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).